moved auth state to core module
