+++ src/app/app.module.ts
@@ -14,7 +14,7 @@ import { NgxSpinnerModule } from 'ngx-spinner';
 import { AppComponent } from './app.component';
 import { environment } from 'src/environments/environment';
 import { appReducer } from './core/store/app.state';
-import { AuthEffects } from './auth/state/auth.effects';
+import { AuthEffects } from './core/store/auth/auth.effects';
 import { CoreModule } from './core/core.module';
 import { TaskEffects } from './core/store/todo/task.effects';
 import { NgbActiveModal, NgbModule } from '@ng-bootstrap/ng-bootstrap';
