+++ src/app/auth/components/forgot-password/forgot-password.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { PasswordResetRequestPayload } from 'src/app/core/payloads/auth.payloads';
 import { AppState } from 'src/app/core/store/app.state';
-import { passwordResetRequested } from '../../state/auth.actions';
+import { passwordResetRequested } from '../../../core/store/auth/auth.actions';
 
 @Component({
   selector: 'app-forgot-password',
