+++ src/app/auth/guards/auth.guard.ts
@@ -10,7 +10,7 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { AppState } from 'src/app/core/store/app.state';
-import { isAuthenticated } from '../state/auth.selectors';
+import { isAuthenticated } from '../../core/store/auth/auth.selectors';
 
 @Injectable({
   providedIn: 'root',
