+++ src/app/core/store/auth/auth.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { User } from '../../core/models/user.model';
+import { User } from '../../models/user.model';
 
 export const loginRequest = createAction(
   '[Login Page] Login Request',
