--- src/app/home/components/filter-order-status/status-filter.component.html
@@ -1,15 +1,0 @@
-<div class="status-filter">
-  <label for="filterStatus">Filter by:</label>
-  <select name="filterStatus" id="filterStatus" class="form-control">
-    <option value="Not started">Not started</option>
-    <option value="In progress">In progress</option>
-    <option value="Completed">Completed</option>
-  </select>
-  <label for="orderByStatus">Order by:</label>
-  <select name="orderByStatus" id="orderByStatus" class="form-control">
-    <option value="asc">Ascending</option>
-    <option value="desc">Description</option>
-  </select>
-
-  <button class="btn btn-outline-secondary ml-2">Go</button>
-</div>
