--- src/app/home/components/filter-order-status/status-filter.component.spec.ts
@@ -1,24 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { StatusFilterComponent } from './status-filter.component';
-
-describe('FilterOrderStatusComponent', () => {
-  let component: StatusFilterComponent;
-  let fixture: ComponentFixture<StatusFilterComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [StatusFilterComponent],
-    }).compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(StatusFilterComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
