--- src/app/home/components/filter-order-status/status-filter.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { Store } from '@ngrx/store';
-import { AppState } from 'src/app/core/store/app.state';
-
-@Component({
-  selector: 'app-status-filter',
-  templateUrl: './status-filter.component.html',
-  styleUrls: ['./status-filter.component.scss'],
-})
-export class StatusFilterComponent implements OnInit {
-  constructor(private store: Store<AppState>) {}
-
-  ngOnInit(): void {}
-
-  onFilterOrderStatus(): void {}
-}
