+++ src/app/shared/layout/site-header/site-header.component.ts
@@ -1,8 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
-import { signoutRequest } from 'src/app/auth/state/auth.actions';
-import { getUser, isAuthenticated } from 'src/app/auth/state/auth.selectors';
+import { signoutRequest } from 'src/app/core/store/auth/auth.actions';
+import {
+  getUser,
+  isAuthenticated,
+} from 'src/app/core/store/auth/auth.selectors';
 import { User } from 'src/app/core/models/user.model';
 import { AppState } from 'src/app/core/store/app.state';
